requirements.txt
Flask==3.0.3
passlib==1.7.4
PyMySQL==1.1.0
python-dotenv==1.0.1
